refacto: modularization of downloaders.ts for easier maintenance
