+++ server/downloaders/index.ts
@@ -1,0 +1,8 @@
+export * from "./types.js";
+export { DownloaderManager } from "./manager.js";
+export { TransmissionClient } from "./transmission.js";
+export { RTorrentClient } from "./rtorrent.js";
+export { QBittorrentClient } from "./qbittorrent.js";
+export { SABnzbdClient } from "./sabnzbd.js";
+export { NZBGetClient } from "./nzbget.js";
+export { SynologyDownloadStationClient } from "./synology.js";
